# docs/swagger.yaml
definitions:
...
  news.Comment:
    properties:
      image:
        type: string
      message:
        type: string
      name:
        type: string
      time_ago:
        type: string
    type: object
  news.News:
    properties:
      _id:
        type: string
      base:
        $ref: '#/definitions/support.Base'
      caption:
        type: string
      code:
        type: string
      comments:
        items:
          $ref: '#/definitions/news.Comment'
        type: array
      content:
        type: string
      exclusive:
        type: boolean
      featured:
        type: boolean
      meta:
        type: string
      name:
        type: string
      picture:
        type: string
      sections:
        items:
          $ref: '#/definitions/news.Section'
        type: array
      sport:
        type: string
      time_ago:
        type: string
      title:
        type: string
      trending:
        type: boolean
      url:
        type: string
    type: object
  news.Section:
    properties:
      content:
        type: string
      highlight:
        type: boolean
      image:
        type: string
      name:
        type: string
    type: object
...
  support.Base:
    properties:
      created_At:
        type: string
      delete_At:
        type: string
      updated_At:
        type: string
    type: object
  support.HttpError:
    properties:
      code:
        type: integer
      errors:
        type: string
      message:
        type: string
      results:
        type: boolean
    type: object
...
info:
  contact: {}
paths:
...
  /api/news:
    delete:
      consumes:
      - application/json
      description: Delte item
      parameters:
      - description: code
        in: query
        name: code
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/news.News'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/support.HttpError'
      summary: Delete a news
      tags:
      - news
    get:
      consumes:
      - application/json
      description: Get item
      parameters:
      - description: code
        in: query
        name: code
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/news.News'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/support.HttpError'
      summary: Get a news
      tags:
      - news
    post:
      consumes:
      - application/json
      description: Create a new news item
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/news.News'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/support.HttpError'
      summary: Create a news
      tags:
      - news
    put:
      consumes:
      - application/json
      description: Update an new news item
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/news.News'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/support.HttpError'
      summary: Update an news
      tags:
      - news
...
  /news:
    get:
      consumes:
      - application/json
      description: Get  BY Category item
      parameters:
      - description: code
        in: query
        name: code
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/news.News'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/support.HttpError'
      summary: Get  BY Category a news
      tags:
      - news
...
swagger: "2.0"